+++ module-apps/windows/NoEvents.cpp
@@ -2,6 +2,9 @@
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include "NoEvents.hpp"
+#include "Dialog.hpp"
+#include "DialogMetadataMessage.hpp"
+#include "log/log.hpp"
 #include <service-appmgr/ApplicationManager.hpp>
 #include <i18/i18.hpp>
 #include <module-services/service-db/messages/QueryMessage.hpp>
@@ -26,83 +29,20 @@ namespace style
     } // namespace cross
 } // namespace style
 
-NoEvents::NoEvents(app::Application *app, const std::string &name, const NoEvents::Meta &meta)
-    : gui::AppWindow(app, name), meta(meta)
+NoEvents::NoEvents(app::Application *app, const std::string &name) : gui::Dialog(app, name)
 {}
 
-void NoEvents::buildInterface()
+void NoEvents::onBeforeShow(ShowMode mode, SwitchData *data)
 {
-    AppWindow::buildInterface();
-
-    topBar->setActive(gui::TopBar::Elements::TIME, true);
-    bottomBar->setActive(gui::BottomBar::Side::RIGHT, true);
-    bottomBar->setText(gui::BottomBar::Side::RIGHT, utils::localize.get(style::strings::common::back));
-
-    auto app = dynamic_cast<app::ApplicationCalendar *>(application);
-    assert(app != nullptr);
-    if (app->getEquivalentToEmptyWindow() == EquivalentWindow::AllEventsWindow) {
-        bottomBar->setText(gui::BottomBar::Side::LEFT, utils::localize.get("app_calendar_bar_month"));
+    Dialog::onBeforeShow(mode, data);
+    auto metadata = dynamic_cast<DialogMetadataMessage *>(data);
+    if (metadata != nullptr) {
+        auto foo      = metadata->get().action;
+        inputCallback = [foo](Item &, const InputEvent &inputEvent) -> bool {
+            if (foo && inputEvent.isShortPress() && inputEvent.is(KeyCode::KEY_LEFT)) {
+                return foo();
+            }
+            return false;
+        };
     }
-
-    setTitle(meta.title);
-    leftArrowImage = new gui::Image(this, style::arrow::x, style::arrow::y, 0, 0, "arrow_left");
-    newEventImage  = new gui::Image(this, style::cross::x, style::cross::y, 0, 0, "cross");
-
-    emptyListIcon = new gui::Icon(this,
-                                  0,
-                                  style::header::height,
-                                  style::window_width,
-                                  style::window_height - style::header::height - style::footer::height,
-                                  meta.icon,
-                                  utils::localize.get(meta.text));
-}
-
-void NoEvents::update(const Meta &meta)
-{
-    this->meta = meta;
-    rebuild();
-}
-
-void NoEvents::rebuild()
-{
-    AppWindow::destroyInterface();
-    erase();
-    buildInterface();
-}
-
-bool NoEvents::onInput(const gui::InputEvent &inputEvent)
-{
-    auto app = dynamic_cast<app::ApplicationCalendar *>(application);
-    assert(app != nullptr);
-
-    if (inputEvent.keyCode == gui::KeyCode::KEY_RF && inputEvent.state == gui::InputEvent::State::keyReleasedShort) {
-        if (app->getEquivalentToEmptyWindow() == EquivalentWindow::DayEventsWindow) {
-            app->returnToPreviousWindow();
-        }
-        else if (app->getEquivalentToEmptyWindow() == EquivalentWindow::AllEventsWindow) {
-            LOG_DEBUG("Switch to desktop");
-            sapm::ApplicationManager::messageSwitchPreviousApplication(application);
-        }
-    }
-
-    if (AppWindow::onInput(inputEvent)) {
-        return true;
-    }
-
-    if (!inputEvent.isShortPress()) {
-        return false;
-    }
-
-    if (inputEvent.keyCode == gui::KeyCode::KEY_LEFT) {
-        return meta.action();
-    }
-
-    if (inputEvent.keyCode == gui::KeyCode::KEY_LF &&
-        app->getEquivalentToEmptyWindow() == EquivalentWindow::AllEventsWindow) {
-        application->switchWindow(gui::name::window::main_window);
-        LOG_DEBUG("Switch to month view - main window");
-        return true;
-    }
-
-    return false;
 }